Don't make it intrusive into stdexec
